adjust mui settings with layout
