+++ src/app/components/RegisterForm.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { Box, Button, Checkbox, Container, FormControlLabel, Grid, TextField, Typography } from '@mui/material';
 import Image from 'next/image';
