+++ src/app/layout.tsx
@@ -1,6 +1,9 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
+
 import './styles.css';
+import StyledComponentsRegistry from './lib/registry';
+import ThemeProvider from './theme-provider';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -16,7 +19,11 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <StyledComponentsRegistry>
+          <ThemeProvider>{children}</ThemeProvider>
+        </StyledComponentsRegistry>
+      </body>
     </html>
   );
 }
