+++ src/app/register/page.tsx
@@ -1,3 +1,4 @@
+'use client';
 import RegisterForm from '../components/RegisterForm';
 
 const Register = () => {
