--- src/app/theme.ts
@@ -1,18 +1,0 @@
-import { createTheme } from '@mui/material/styles';
-
-const theme = createTheme({
-  typography: {
-    fontFamily: 'IBM Plex Sans KR, sans-serif',
-  },
-  palette: {
-    primary: {
-      main: '#035036',
-      light: '#AFC6AA', // 원하는 색상 코드로 변경
-    },
-    secondary: {
-      main: '#AFC6AA',
-    },
-  },
-});
-
-export default theme;
